requirements.txt
fastapi
uvicorn
crewai
openai
langchain
langchain-huggingface
duckduckgo-search
python-dotenv
pydantic
pysqlite3-binary